requirements.txt
flask
flask-socketio
RPi.GPIO
pillow
